Restructured frontend
