+++ frontend/ByteMe/app/(header)/header.jsx
@@ -1,12 +1,2 @@
-import { View, Text } from 'react-native'
-import React from 'react'
-
-const header = () => {
-  return (
-    <View>
-      <Text>header</Text>
-    </View>
-  )
-}
-
-export default header
+import HeaderScreen from "@/screens/HeaderScreen";
+export default HeaderScreen;
