+++ frontend/ByteMe/app/(header)/profile.jsx
@@ -1,12 +1,2 @@
-import { View, Text } from 'react-native'
-import React from 'react'
-
-const Profile = () => {
-  return (
-    <View>
-      <h1>Profile</h1>
-    </View>
-  )
-}
-
-export default Profile
+import ProfileScreen from "@/screens/ProfileScreen";
+export default ProfileScreen;
