+++ frontend/ByteMe/app/(start)/forgotPassword.jsx
@@ -1,14 +1,2 @@
-import { StyleSheet, Text, View } from 'react-native'
-import React from 'react'
-
-const ForgotPassword = () => {
-  return (
-    <View>
-      <Text>ForgotPassword</Text>
-    </View>
-  )
-}
-
-export default ForgotPassword
-
-const styles = StyleSheet.create({})
+import ForgotPasswordScreen from "@/screens/ForgetPasswordScreen";
+export default ForgotPasswordScreen;
