+++ frontend/ByteMe/app/(tabs)/calendar.jsx
@@ -1,12 +1,2 @@
-import { View, Text } from 'react-native'
-import React from 'react'
-
-const Calendar = () => {
-  return (
-    <View>
-      <Text>Calendar</Text>
-    </View>
-  )
-}
-
-export default Calendar
+import CalendarScreen from "@/screens/CalendarScreen";
+export default CalendarScreen;
