+++ frontend/ByteMe/app/(tabs)/grocery.jsx
@@ -1,12 +1,2 @@
-import { View, Text } from 'react-native'
-import React from 'react'
-
-const Grocery = () => {
-  return (
-    <View>
-      <Text>Grocery</Text>
-    </View>
-  )
-}
-
-export default Grocery
+import GroceryScreen from "@/screens/GroceryScreen";
+export default GroceryScreen;
