+++ frontend/ByteMe/app/(tabs)/pantry.jsx
@@ -1,12 +1,2 @@
-import { View, Text } from 'react-native'
-import React from 'react'
-
-const Pantry = () => {
-  return (
-    <View>
-      <Text>Pantry</Text>
-    </View>
-  )
-}
-
-export default Pantry
+import PantryScreen from "@/screens/PantryScreen";
+export default PantryScreen;
