+++ frontend/ByteMe/app/(tabs)/savedrecipes.jsx
@@ -1,12 +1,2 @@
-import { View, Text } from 'react-native'
-import React from 'react'
-
-const Recipes = () => {
-  return (
-    <View>
-      <Text>Recipes</Text>
-    </View>
-  )
-}
-
-export default Recipes
+import SavedRecipesScreen from "@/screens/SavedRecipesScreen";
+export default SavedRecipesScreen;
