+++ frontend/ByteMe/screens/CalendarScreen.jsx
@@ -1,0 +1,12 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+const Calendar = () => {
+  return (
+    <View>
+      <Text>Calendar</Text>
+    </View>
+  )
+}
+
+export default Calendar
