+++ frontend/ByteMe/screens/ForgetPasswordScreen.jsx
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const ForgotPassword = () => {
+  return (
+    <View>
+      <Text>ForgotPassword</Text>
+    </View>
+  )
+}
+
+export default ForgotPassword
+
+const styles = StyleSheet.create({})
