+++ frontend/ByteMe/screens/GroceryScreen.jsx
@@ -1,0 +1,12 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+const Grocery = () => {
+  return (
+    <View>
+      <Text>Grocery</Text>
+    </View>
+  )
+}
+
+export default Grocery
