+++ frontend/ByteMe/screens/HeaderScreen.jsx
@@ -1,0 +1,12 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+const header = () => {
+  return (
+    <View>
+      <Text>header</Text>
+    </View>
+  )
+}
+
+export default header
