+++ frontend/ByteMe/screens/PantryScreen.jsx
@@ -1,0 +1,12 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+const Pantry = () => {
+  return (
+    <View>
+      <Text>Pantry</Text>
+    </View>
+  )
+}
+
+export default Pantry
