+++ frontend/ByteMe/screens/ProfileScreen.jsx
@@ -1,0 +1,12 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+const Profile = () => {
+  return (
+    <View>
+      <h1>Profile</h1>
+    </View>
+  )
+}
+
+export default Profile
