+++ frontend/ByteMe/screens/SavedRecipesScreen.jsx
@@ -1,0 +1,12 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+const Recipes = () => {
+  return (
+    <View>
+      <Text>Recipes</Text>
+    </View>
+  )
+}
+
+export default Recipes
